fext: mexc v1

+++ model/stock.go
@@ -22,7 +22,7 @@ type StockData struct {
 	High      float64   `json:"high"`
 	Low       float64   `json:"low"`
 	Close     float64   `json:"close"`
-	Volume    int64     `json:"volume"`
+	Volume    float64   `json:"volume"`
 	OpenTime  time.Time `json:"open_time"`
 	CloseTime time.Time `json:"close_time"`
 	QuoteVol  float64   `json:"quote_volume"`
